Moved inject/extract test into fasthttpcarrier tests
